+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 CHANGELOG
 =========
 
+6.1.0
+ - Added the Bumper Page as another option to overlap to each ad (video, interstitial, banner, app wall)
+ - Separated the parental gate into a separate library (and done some refactoring)
+
 6.0.4
  - Add a null pointer exception check to SABannerAd to guard against the case when an ad started playing but is closed before "Web_Loaded" gets called